+++ CHANGELOG.md
@@ -22,9 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add voting info to labels selector [#840](https://github.com/PublicMapping/districtbuilder/pull/840)
 - Add evaluate mode metric view for competitiveness [#824](https://github.com/PublicMapping/districtbuilder/pull/824)
 - Add expandable metrics viewer with ability to pin metrics to sidebar [#827](https://github.com/PublicMapping/districtbuilder/pull/827)
-- Update population deviation helper text [#848](https://github.com/PublicMapping/districtbuilder/pull/848)
 - Add user id and IP address to rollbar server side error logging [#841](https://github.com/PublicMapping/districtbuilder/pull/841)
-
+- Add majority race metric to project sidebar [#853](https://github.com/PublicMapping/districtbuilder/pull/853)
 
 ### Changed
 
@@ -35,6 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rename Support menu to Resources [#787](https://github.com/PublicMapping/districtbuilder/pull/787)
 - Switched race demographic colors to a palette with fewer conflicts to our other color palettes [#795](https://github.com/PublicMapping/districtbuilder/pull/795)
 - Tweaks to org report downloads [#832](https://github.com/PublicMapping/districtbuilder/pull/832)
+- Update population deviation helper text [#848](https://github.com/PublicMapping/districtbuilder/pull/848)
+
 
 ### Fixed